01MalayalamLearningApp/ar-language-app/backend/requirements.txt
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
python-multipart>=0.0.5
pillow==10.0.0
google-generativeai>=0.3.0
openai-whisper==20231117
soundfile==0.12.1
torch==2.2.1
numpy>=1.24.3
gTTS==2.5.1
ffmpeg-python==0.2.0
whisper
soundfile
pydantic-settings>=2.0.0